Init file and support to store last imported project (for easy debug)
Support for the init file (anm.js), which configures the namespace, manages the way of player initialization process, prepares and stores some private global configuration
This file is now also included into bundles.
Initialization blocks in sources are now look much more readable thanks to this file and now they contain just a few lines.
Now it is possible for developer to configure all player instances using some global variables, this way is now used to decide if log an imported project or not.