postpone Audio.play when it's not yet ready
This fixes the sutiation when the playback starts before
`ResourceManager` hands off the audio data to the object, which
happened when the audio was loaded from cache - audio’s `loader()`
function was never called so it wasn’t initialized correctly for the
second time. This is now fixed.

This is a bit of a dirty solution, but a good one requires changing how
the RM works and takes much more time, so I’m fixing it as it is now.